+++ examples/solid-ts/src/errors/404.tsx
@@ -1,8 +1,0 @@
-export default function NotFound() {
-  return (
-    <section>
-      <h1>404: Not Found</h1>
-      <p>It's gone 😞</p>
-    </section>
-  )
-}