+++ packages/utilities/dom-query/src/platform.ts
@@ -5,14 +5,19 @@ export function getPlatform() {
   return agent?.platform ?? navigator.platform
 }
 
+if (isDom()) console.log(getPlatform())
+
 const pt = (v: RegExp) => isDom() && v.test(getPlatform())
 const ua = (v: RegExp) => isDom() && v.test(navigator.userAgent)
 const vn = (v: RegExp) => isDom() && v.test(navigator.vendor)
 
 export const isTouchDevice = () => isDom() && !!navigator.maxTouchPoints
-export const isMac = () => pt(/^Mac/) && !isTouchDevice()
+export const isMac = () => pt(/^Mac/)
 export const isSafari = () => isApple() && vn(/apple/i)
 export const isFirefox = () => ua(/firefox\//i)
 export const isApple = () => pt(/mac|iphone|ipad|ipod/i)
 export const isIos = () => pt(/iP(hone|ad|od)|iOS/)
 export const isWebKit = () => ua(/AppleWebKit/)
+
+export const isModKey = (event: Pick<KeyboardEvent, "metaKey" | "ctrlKey">) =>
+  isApple() ? event.metaKey : event.ctrlKey